requirements.txt
fastapi
uvicorn
requests
mkdocs
python-dotenv
streamlit
requests
groq
pydantic


